FIXED: Several warnings ADDED: Methods which get filenames as parameters exist now in two variants (where possible)        1. version taking const char*, takes ASCII or UTF-8 on Linux        2. version taking const wchar_t* takes Unicode on every platform ADDED: PdfString can now be initialized from wchar_t*
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@880 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfFontTTFSubset.h
@@ -183,9 +183,9 @@ inline unsigned long PdfFontTTFSubset::GetSize() const
 inline void PdfFontTTFSubset::AddCharacter( pdf_utf16be nCharCode )
 {
 #ifdef PODOFO_IS_LITTLE_ENDIAN
-    this->AddGlyph( m_pMetrics->GetGlyphId( ((nCharCode & 0xff00) >> 8) | ((nCharCode & 0xff) << 8) ) );
+    this->AddGlyph( static_cast<unsigned short>(m_pMetrics->GetGlyphId( ((nCharCode & 0xff00) >> 8) | ((nCharCode & 0xff) << 8) )) );
 #else
-    this->AddGlyph( m_pMetrics->GetGlyphId( nCharCode ) );
+    this->AddGlyph( static_cast<unsigned short>(m_pMetrics->GetGlyphId( nCharCode )) );
 #endif // PODOFO_IS_LITTLE_ENDIAN
 }
